+++ Sources/SpeziMedication/ModifyMedication/EditFrequency.swift
@@ -1,0 +1,43 @@
+//
+// This source file is part of the Stanford Spezi open-source project
+//
+// SPDX-FileCopyrightText: 2022 Stanford University and the project authors (see CONTRIBUTORS.md)
+//
+// SPDX-License-Identifier: MIT
+//
+
+import SwiftUI
+
+
+struct EditFrequency: View {
+    @Binding private var frequency: Frequency
+    @State private var showFrequencySheet: Bool = false
+    
+    
+    var body: some View {
+        Section {
+            Button(
+                action: {
+                    showFrequencySheet.toggle()
+                },
+                label: {
+                    HStack {
+                        Text("Frequency")
+                            .foregroundStyle(Color.primary)
+                        Spacer()
+                        Text(frequency.description)
+                            .foregroundStyle(Color.accentColor)
+                    }
+                }
+            )
+        }
+            .sheet(isPresented: $showFrequencySheet) {
+                ScheduleFrequencyView(frequency: $frequency)
+            }
+    }
+    
+    
+    init(frequency: Binding<Frequency>) {
+        self._frequency = frequency
+    }
+}
