+++ Sources/SpeziMedication/ModifyMedication/EditMedication.swift
@@ -15,6 +15,7 @@ struct EditMedication<MI: MedicationInstance>: View {
     @Environment(InternalMedicationSettingsViewModel<MI>.self) private var viewModel
     
     @State private var dosage: MI.InstanceDosage
+    @State private var schedule: Schedule
     
     private let medicationInstanceId: MI.ID
         
@@ -27,7 +28,16 @@ struct EditMedication<MI: MedicationInstance>: View {
         if let medicationInstance {
             VStack {
                 Form {
-                    EditDosageSection<MI>(dosage: $dosage, medication: medicationInstance.type, initialDosage: dosage)
+                    Section(String(localized: "Dosage", bundle: .module)) {
+                        EditDosage<MI>(dosage: $dosage, medication: medicationInstance.type, initialDosage: dosage)
+                            .labelsHidden()
+                    }
+                    Section(String(localized: "Schedule", bundle: .module)) {
+                        EditFrequency(frequency: $schedule.frequency)
+                    }
+                    Section(String(localized: "Schedule Times", bundle: .module)) {
+                        EditScheduleTime(times: $schedule.times)
+                    }
                     Section {
                         Button(String(localized: "Delete", bundle: .module), role: .destructive) {
                             viewModel.medicationInstances.remove(medicationInstance)
@@ -49,6 +59,18 @@ struct EditMedication<MI: MedicationInstance>: View {
                     
                     viewModel.medicationInstances.insert(modifiedMedication)
                 }
+                .onChange(of: schedule) {
+                    guard medicationInstance.schedule != schedule else {
+                        return
+                    }
+                    
+                    viewModel.medicationInstances.remove(medicationInstance)
+                    
+                    var modifiedMedication = medicationInstance
+                    modifiedMedication.schedule = schedule
+                    
+                    viewModel.medicationInstances.insert(modifiedMedication)
+                }
         } else {
             ZStack {
                 Color(.systemGroupedBackground).ignoresSafeArea()
@@ -60,9 +82,11 @@ struct EditMedication<MI: MedicationInstance>: View {
     
     init(
         medicationInstance medicationInstanceId: MI.ID,
-        initialDosage: MI.InstanceDosage
+        initialDosage: MI.InstanceDosage,
+        initialSchedule: Schedule
     ) {
         self.medicationInstanceId = medicationInstanceId
         self._dosage = State(initialValue: initialDosage)
+        self._schedule = State(initialValue: initialSchedule)
     }
 }
